# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Ordenes",
        "description": "Documentación de la API para gestionar órdenes.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener todos los productos",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida satisfactoriamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "status": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "response": {
                                                    "properties": {
                                                        "message": {
                                                            "type": "string",
                                                            "example": "Productos obtenidos satisfactoriamente"
                                                        },
                                                        "body": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "Producto 1"
                                                                    },
                                                                    "price": {
                                                                        "type": "number",
                                                                        "format": "float",
                                                                        "example": 23.4
                                                                    },
                                                                    "category": {
                                                                        "type": "string",
                                                                        "example": "shoe"
                                                                    },
                                                                    "quantity": {
                                                                        "type": "integer",
                                                                        "example": 5
                                                                    },
                                                                    "created_at": {
                                                                        "type": "string",
                                                                        "format": "date-time",
                                                                        "example": "2024-09-11T20:48:45.000000Z"
                                                                    },
                                                                    "updated_at": {
                                                                        "type": "string",
                                                                        "format": "date-time",
                                                                        "example": "2024-09-11T20:48:45.000000Z"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud inválida"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 23.4
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "shoe"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto creado satisfactoriamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "status": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "response": {
                                                    "properties": {
                                                        "message": {
                                                            "type": "string",
                                                            "example": "Producto creado satisfactoriamente"
                                                        },
                                                        "body": {
                                                            "type": "integer",
                                                            "example": 319
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud inválida"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
...
    },
...
}